+++ Sources/BungoManifest/Protocols/BungoManifestSQLClient/BungoManifestSQLClient.swift
@@ -3,5 +3,5 @@ import Foundation
 public protocol BungoManifestSQLClient {
     init(at path: String) throws
 
-    func select(column: String, from table: String, where idColumn: String, is id: Int64) throws -> Data?
+    func select(column: String, from table: String, where idColumn: String, is id: Int32) throws -> Data?
 }
